+++ src/goldsberry/endpoints/player/__init__.py
@@ -6,6 +6,7 @@ from .career_stats import (
     get_career_stats,
     get_career_stats_async,
 )
+from .game_logs import GameLogsEndpoint, get_game_logs, get_game_logs_async
 
 __all__ = [
     "PlayerListEndpoint",
@@ -14,4 +15,7 @@ __all__ = [
     "CareerStatsEndpoint",
     "get_career_stats",
     "get_career_stats_async",
+    "GameLogsEndpoint",
+    "get_game_logs",
+    "get_game_logs_async",
 ]