+++ src/goldsberry/models/player/__init__.py
@@ -2,5 +2,12 @@
 
 from .player_list import PlayerInfo
 from .career_stats import SeasonTotals, CareerTotals, PlayerCareerStats
+from .game_logs import GameLog
 
-__all__ = ["PlayerInfo", "SeasonTotals", "CareerTotals", "PlayerCareerStats"]
+__all__ = [
+    "PlayerInfo",
+    "SeasonTotals",
+    "CareerTotals",
+    "PlayerCareerStats",
+    "GameLog",
+]